- Back to old non-relay world - Moved supplementary gulpfiles to gfiles directory - removed unused deps
