+++ package.json
@@ -21,7 +21,6 @@
     "graceful-fs": "^4.1.3",
     "graphql": "^0.5.0",
     "lodash": "^4.11.1",
-    "material-ui": "^0.14.4",
     "mime": "^1.3.4",
     "mongodb": "^2.1.16",
     "nconf": "^0.8.4",
